working unit test

+++ src/bu_solver_mlp.h
@@ -13,7 +13,7 @@ typedef std::map<Nuc, double> CompMap;
 
 /// For the moment, this is just an adaptation of the CLASS-MLP Model to CYCLUS
 /// @code
-/// [1] B. Leniau, B. Mouginota, N. Thiolliere et al. "A neural network approach
+/// [1] B. Leniau, B. Mouginot, N. Thiolliere et al. "A neural network approach
 /// for burn-up calculation and its application to the dynamic fuel cycle code
 /// CLASS" Annals of Nuclear Energy, Volume 81, July 2015
 /// @endcode
@@ -31,12 +31,12 @@ namespace cybam {
 
 
         // Lauch the MLP (using TMVA) to predict the requeirt fissil enrichment (according to the compisition of fissil and fertil stream and the targeted burnup)
-        double GetEnrichment(cyclus::Composition::Ptr c_fissil,
+        float GetEnrichment(cyclus::Composition::Ptr c_fissil,
                            cyclus::Composition::Ptr c_fertil,
                            double BurnUp);
 
         // liner dicchotomy to determine the Burn-up reachable by a fuel depending of its composition...
-        double GetBU(cyclus::Composition::Ptr fuel, double eps = 1e-6 );
+        float GetBU(cyclus::Composition::Ptr fuel, double eps = 1e-6 );
 
         std::string str() { return TMVAWeightFile;};
